UserTest presque OK

+++ routes/web.php
@@ -40,13 +40,13 @@ Route::patch('/home', [HomeController::class, 'updatePassword'])->name('updatePa
 Route::get('/admin', [AdminController::class, 'index'])->name('admin')->middleware('admin');
 // Admin videos
 Route::get('/admin/video/edit/{video}', [AdminController::class, 'edit'])->name('admin.edit.video')->middleware('admin');
-Route::patch('/admin/video/edit/{video}', [AdminController::class, 'update'])->name('admin.update');
+Route::patch('/admin/video/edit/{video}', [AdminController::class, 'update'])->name('admin.update')->middleware('admin');
 // Admin users
 Route::get('/admin/user/create', [UserController::class, 'showCreateUserForm'])->name('admin.user.showForm')->middleware('admin');
-Route::post('/admin/user/create', [UserController::class, 'store'])->name('admin.user.store');
+Route::post('/admin/user/create', [UserController::class, 'store'])->name('admin.user.store')->middleware('admin');
 Route::get('/admin/user/edit/{user}', [UserController::class, 'edit'])->name('admin.user.edit')->middleware('admin');
-Route::patch('/admin/user/edit/{user}', [UserController::class, 'update'])->name('admin.user.update');
-Route::delete('/admin/user/delete/{user}', [UserController::class, 'destroy'])->name('admin.user.destroy');
+Route::patch('/admin/user/edit/{user}', [UserController::class, 'update'])->name('admin.user.update')->middleware('admin');
+Route::delete('/admin/user/delete/{user}', [UserController::class, 'destroy'])->name('admin.user.destroy')->middleware('admin');
 
 // Sitemap
 Route::get('/sitemap/videos', [SitemapController::class, 'videos'])->name('sitemap.videos');
